refactor: added try_from block for header record. Moved to header module
