+++ src/headers/mod.rs
@@ -47,3 +47,33 @@ pub fn check_valid_header(block: &Block, header_dir: &str) -> Result<(), BlockHe
 
     Ok(())
 }
+
+#[derive(Serialize, Deserialize)]
+pub struct HeaderRecordWithNumber {
+    pub block_hash: Vec<u8>,
+    pub total_difficulty: Vec<u8>,
+    pub block_number: u64,
+}
+
+impl TryFrom<Block> for HeaderRecordWithNumber {
+    type Error = BlockHeaderError;
+    fn try_from(block: Block) -> Result<Self, Self::Error> {
+        let block_header = match block.header.clone() {
+            Some(header) => header,
+            None => {
+                return Err(BlockHeaderError::MissingHeader);
+            }
+        };
+        let header_record_with_number = HeaderRecordWithNumber {
+            block_hash: block.hash.clone(),
+            total_difficulty: block_header
+                .total_difficulty
+                .as_ref()
+                .ok_or(BlockHeaderError::InvalidTotalDifficulty)?
+                .bytes
+                .clone(),
+            block_number: block.number,
+        };
+        Ok(header_record_with_number)
+    }
+}
